investigating test failure on jenkins
Former-commit-id: 9d69f6a02aef88477594969524656aa6ac3e9e4e

+++ record_classification/src/test/java/uk/ac/standrews/cs/digitising_scotland/record_classification/cli/ConfigurationTest.java
@@ -37,7 +37,9 @@ public class ConfigurationTest {
     @After
     public void tearDown() throws Exception {
 
-        FileManipulation.deleteDirectory(Configuration.CLI_HOME);
+        if (Files.isDirectory(Configuration.CLI_HOME)) {
+            FileManipulation.deleteDirectory(Configuration.CLI_HOME);
+        }
     }
 
     @Test
